# Vector3dvector and other vector Eigen bindings speed up