--- a/src/UnitTest/Python/readme.md
+++ b/src/UnitTest/Python/readme.md
@@ -1,0 +1,20 @@
+# Python Unit Test
+
+Install:
+
+```bash
+pip install pytest
+```
+
+Run:
+
+```bash
+# Picks up all files named "test_*.py", and runs all functions named "test_*"
+pytest
+
+# Or point to a test directory, e.g.
+pytest /path/to/root/test/directory
+
+# Use -s to show stdout
+pytest -s
+```