@qianyizh 2018-11-02
This tries to address #403