@yxlao 2018-11-02
Update: squeezed in another optimization ed122ac with direct memory mapping, 30% more speed up:
```
open3d.Vector3dVector: (2000000, 3)
open3d -> numpy: 0.017295s
numpy -> open3d: 0.000016s

open3d.Vector3iVector: (2000000, 3)
open3d -> numpy: 0.009439s
numpy -> open3d: 0.002394s

open3d.Vector2iVector: (2000000, 3)
open3d -> numpy: 0.004427s
numpy -> open3d: 0.001198s
```